Every function in its own file, part 7 [#CLICKHOUSE-2]
