+++ dbms/src/Functions/bitRotateLeft.cpp
@@ -1,0 +1,41 @@
+#include <Functions/FunctionFactory.h>
+#include <Functions/FunctionBinaryArithmetic.h>
+
+namespace DB
+{
+
+template <typename A, typename B>
+struct BitRotateLeftImpl
+{
+    using ResultType = typename NumberTraits::ResultOfBit<A, B>::Type;
+
+    template <typename Result = ResultType>
+    static inline Result apply(A a, B b)
+    {
+        return (static_cast<Result>(a) << static_cast<Result>(b))
+            | (static_cast<Result>(a) >> ((sizeof(Result) * 8) - static_cast<Result>(b)));
+    }
+
+#if USE_EMBEDDED_COMPILER
+    static constexpr bool compilable = true;
+
+    static inline llvm::Value * compile(llvm::IRBuilder<> & b, llvm::Value * left, llvm::Value * right, bool)
+    {
+        if (!left->getType()->isIntegerTy())
+            throw Exception("BitRotateLeftImpl expected an integral type", ErrorCodes::LOGICAL_ERROR);
+        auto * size = llvm::ConstantInt::get(left->getType(), left->getType()->getPrimitiveSizeInBits());
+        /// XXX how is this supposed to behave in signed mode?
+        return b.CreateOr(b.CreateShl(left, right), b.CreateLShr(left, b.CreateSub(size, right)));
+    }
+#endif
+};
+
+struct NameBitRotateLeft { static constexpr auto name = "bitRotateLeft"; };
+using FunctionBitRotateLeft = FunctionBinaryArithmetic<BitRotateLeftImpl, NameBitRotateLeft>;
+
+void registerFunctionBitRotateLeft(FunctionFactory & factory)
+{
+    factory.registerFunction<FunctionBitRotateLeft>();
+}
+
+}
