+++ dbms/src/Functions/bitTest.cpp
@@ -1,0 +1,31 @@
+#include <Functions/FunctionFactory.h>
+#include <Functions/FunctionBinaryArithmetic.h>
+
+namespace DB
+{
+
+template <typename A, typename B>
+struct BitTestImpl
+{
+    using ResultType = UInt8;
+
+    template <typename Result = ResultType>
+    static inline Result apply(A a, B b)
+    {
+        return (typename NumberTraits::ToInteger<A>::Type(a) >> typename NumberTraits::ToInteger<B>::Type(b)) & 1;
+    }
+
+#if USE_EMBEDDED_COMPILER
+    static constexpr bool compilable = false; /// TODO
+#endif
+};
+
+struct NameBitTest { static constexpr auto name = "bitTest"; };
+using FunctionBitTest = FunctionBinaryArithmetic<BitTestImpl, NameBitTest>;
+
+void registerFunctionBitTest(FunctionFactory & factory)
+{
+    factory.registerFunction<FunctionBitTest>();
+}
+
+}
