+++ dbms/src/Functions/lcm.cpp
@@ -1,0 +1,36 @@
+#include <Functions/FunctionFactory.h>
+#include <Functions/FunctionBinaryArithmetic.h>
+#include <boost/integer/common_factor.hpp>
+
+namespace DB
+{
+
+template <typename A, typename B>
+struct LCMImpl
+{
+    using ResultType = typename NumberTraits::ResultOfAdditionMultiplication<A, B>::Type;
+
+    template <typename Result = ResultType>
+    static inline Result apply(A a, B b)
+    {
+        throwIfDivisionLeadsToFPE(typename NumberTraits::ToInteger<A>::Type(a), typename NumberTraits::ToInteger<B>::Type(b));
+        throwIfDivisionLeadsToFPE(typename NumberTraits::ToInteger<B>::Type(b), typename NumberTraits::ToInteger<A>::Type(a));
+        return boost::integer::lcm(
+            typename NumberTraits::ToInteger<Result>::Type(a),
+            typename NumberTraits::ToInteger<Result>::Type(b));
+    }
+
+#if USE_EMBEDDED_COMPILER
+    static constexpr bool compilable = false; /// exceptions (and a non-trivial algorithm)
+#endif
+};
+
+struct NameLCM { static constexpr auto name = "lcm"; };
+using FunctionLCM = FunctionBinaryArithmetic<LCMImpl, NameLCM, false>;
+
+void registerFunctionLCM(FunctionFactory & factory)
+{
+    factory.registerFunction<FunctionLCM>();
+}
+
+}
